my-siwe-backend/requirements.txt
abnf==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backports.tarfile==1.2.0
bitarray==3.5.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
cleo==2.1.0
click==8.2.1
crashtest==0.4.1
cryptography==45.0.5
cytoolz==1.0.1
distlib==0.3.9
dnspython==2.7.0
dulwich==0.22.8
ecdsa==0.19.1
email_validator==2.2.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.116.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.2
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
frozenlist==1.7.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.7.0
msgpack==1.1.1
multidict==6.6.3
packaging==25.0
parsimonious==0.10.0
pbs-installer==2025.7.8
pkginfo==1.12.1.2
platformdirs==4.3.8
poetry-core==2.1.3
propcache==0.3.2
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyproject_hooks==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pyunormalize==16.0.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.5.1
rlp==4.1.0
rsa==4.9.1
SecretStorage==3.3.3
sentry-sdk==2.32.0
shellingham==1.5.4
siwe==4.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tomlkit==0.13.3
toolz==1.0.0
trove-classifiers==2025.5.9.12
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
web3==7.12.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
